@@ -8,7 +8,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
 # followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
+# Both the version and the   number may be overridden in flutter
 # build by specifying --build-name and --build-number, respectively.
 # In Android, build-name is used as versionName while build-number used as versionCode.
 # Read more about Android versioning at https://developer.android.com/studio/publish/versioning
@@ -18,8 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-
+  sdk: ">=2.17.6 <3.0.0"
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
 # consider running `flutter pub upgrade --major-versions`. Alternatively,
@@ -46,16 +45,21 @@ dependencies:
   flutter_widget_from_html: ^0.8.5
   cupertino_stepper: ^0.2.1
   enhance_stepper: ^0.1.1
-  sms_autofill: ^2.2.0
+  telephony: ^0.2.0
+  otp_text_field: ^1.1.3
   dropdown_button2: ^1.6.3
   file_picker: ^5.0.0
   percent_indicator: ^4.2.2
-  scrollable_positioned_list: ^0.2.0-nullsafety.0
+  scrollable_positioned_list: ^0.3.4
   expandable_page_view: ^1.0.17
   encrypt: ^5.0.1
   intl: ^0.17.0
   device_info_plus: ^4.1.2
   qr_flutter: ^4.0.0
+  clipboard: ^0.1.3
+  #path_provider: ^2.0.11
+  permission_handler: ^10.1.0
+  external_path: ^1.0.3
 
 dev_dependencies:
   flutter_test:
