@@ -6,8 +6,10 @@ import 'package:andapp/common/string_utils.dart';
 import 'package:andapp/di/app_component_base.dart';
 import 'package:andapp/di/shared_preferences.dart';
 import 'package:andapp/model/get_question_answer_list.dart';
+import 'package:andapp/model/submit_answer.dart';
 import 'package:andapp/screen/training/training_result.dart';
 import 'package:andapp/services/api_client.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 class TrainingExamBloc extends BlocBase {
@@ -40,7 +42,7 @@ class TrainingExamBloc extends BlocBase {
     });
   }
 
-  Future submitAnswers(BuildContext context, String trainingType,
+  Future<AnswerData?> submitAnswers(BuildContext context, String trainingType,
       List<Question>? lstQuestions) async {
     List<AnswerList> listAnswer = [];
     for (Question question in lstQuestions ?? []) {
@@ -49,43 +51,43 @@ class TrainingExamBloc extends BlocBase {
       objAnswer.ans = question.attemptedAns;
       listAnswer.add(objAnswer);
     }
-    AppComponentBase.getInstance()
+    var pospId = await AppComponentBase.getInstance()
         ?.getSharedPreference()
-        .getUserDetail(key: SharedPreference().pospId)
-        .then((value) async => await AppComponentBase.getInstance()
-                ?.getApiInterface()
-                .getApiRepository()
-                .submitAnswers(
-                    trainingType: trainingType == StringUtils.generalInsurance
-                        ? ApiClient.trainingTypeGI
-                        : ApiClient.trainingTypeLI,
-                    ansList: listAnswer,
-                    pospId: value)
-                .then((submitAnswer) {
-              print("in : submitAnswer : ${submitAnswer?.messages}");
-              if (submitAnswer != null &&
-                  submitAnswer.resultflag == ApiClient.resultflagSuccess &&
-                  submitAnswer.data != null) {
-                print("if : submitAnswer : ${submitAnswer.messages}");
-                //return submitAnswer;
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) {
-                    return TrainingResult(
-                      title: trainingType,
-                      answerData: submitAnswer.data?.data,
-                    );
-                  }),
-                );
-              } else {
-                CommonToast.getInstance()
-                    ?.displayToast(message: submitAnswer?.messages ?? "Fail");
-                print("else : submitAnswer : ${submitAnswer?.messages}");
-                //return null;
-              }
-            }));
-
-    print("out : submitAnswer is null");
+        .getUserDetail(key: SharedPreference().pospId);
+    if(pospId != null) {
+      var submitAnswer = await AppComponentBase.getInstance()
+          ?.getApiInterface()
+          .getApiRepository()
+          .submitAnswers(
+          trainingType: trainingType == StringUtils.generalInsurance
+              ? ApiClient.trainingTypeGI
+              : ApiClient.trainingTypeLI,
+          ansList: listAnswer,
+          pospId: pospId);
+      if (submitAnswer != null) {
+        if (kDebugMode) {
+          print("in : submitAnswer : ${submitAnswer.messages}");
+        }
+        if (submitAnswer.resultflag == ApiClient.resultflagSuccess &&
+            submitAnswer.data != null) {
+          if (kDebugMode) {
+            print("if : submitAnswer : ${submitAnswer.messages}");
+          }
+        } else {
+          CommonToast.getInstance()
+              ?.displayToast(message: submitAnswer.messages ?? "Fail");
+          if (kDebugMode) {
+            print("else : submitAnswer : ${submitAnswer.messages}");
+          }
+          //return null;
+        }
+        return submitAnswer.data?.data;
+      }
+    }
+    if (kDebugMode) {
+      print("out : submitAnswer is null");
+    }
+    return null;
   }
 
   @override
