@@ -1,15 +1,18 @@
 import 'package:andapp/common/app_theme.dart';
 import 'package:andapp/common/image_utils.dart';
+import 'package:andapp/common/pink_border_button.dart';
 import 'package:andapp/common/string_utils.dart';
 import 'package:andapp/screen/nav_bar.dart';
+import 'package:andapp/screen/training/training_exam.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 class TrainingNavigator extends StatefulWidget {
   final String title;
   final String day;
+  final String pospId;
 
-  const TrainingNavigator({Key? key, required this.title, required this.day})
+  const TrainingNavigator({Key? key, required this.title,required this.pospId, required this.day})
       : super(key: key);
 
   @override
@@ -29,7 +32,7 @@ class _TrainingNavigatorState extends State<TrainingNavigator>
     } else if (widget.day == "2") {
       day1 = true;
       day2 = true;
-    } else if (widget.day == "2") {
+    } else if (widget.day == "3") {
       day1 = true;
       day2 = true;
       day3 = true;
@@ -78,7 +81,7 @@ class _TrainingNavigatorState extends State<TrainingNavigator>
                   mainAxisSize: MainAxisSize.max,
                   children: [
                     Column(
-                        mainAxisAlignment: MainAxisAlignment.start,
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         crossAxisAlignment: CrossAxisAlignment.stretch,
                         mainAxisSize: MainAxisSize.max,
                         children: [
@@ -198,72 +201,96 @@ class _TrainingNavigatorState extends State<TrainingNavigator>
                               )),
                           //const SizedBox(height: 200,),
                         ]),
-                    Align(
-                      alignment: Alignment.bottomCenter,
-                      child: Padding(
-                        padding: const EdgeInsets.symmetric(
-                            horizontal: 48.0, vertical: 16),
-                        child: Column(
-                          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                          children: [
-                            SvgPicture.asset(
-                              SvgImages.trainingResult,
-                              height: 48,
-                              width: 48,
-                              alignment: Alignment.topRight,
-                            ),
-                            const SizedBox(
-                              height: 24,
-                            ),
-                            Text(
-                                widget.day == "3"
-                                    ? StringUtils.trainingDaysSuccessTitle
-                                    : StringUtils.trainingSuccessTitle,
-                                style: const TextStyle(
-                                    fontSize: 17, fontWeight: FontWeight.w500)),
-                            const SizedBox(
-                              height: 20,
-                            ),
-                            widget.day == "1"
-                                ? const Text(
-                                    StringUtils.trainingDay1CompleteSuccess,
-                                    style: TextStyle(
-                                        fontSize: 14,
-                                        fontWeight: FontWeight.w300))
-                                : widget.day == "2"
+                    Column(
+                      mainAxisAlignment: MainAxisAlignment.end,
+                      crossAxisAlignment: CrossAxisAlignment.stretch,
+                      children: [
+                        Align(
+                          alignment: Alignment.bottomCenter,
+                          child: Padding(
+                            padding: const EdgeInsets.symmetric(
+                                horizontal: 48.0, vertical: 16),
+                            child: Column(
+                              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                              children: [
+                                SvgPicture.asset(
+                                  SvgImages.trainingResult,
+                                  height: 48,
+                                  width: 48,
+                                  alignment: Alignment.topRight,
+                                ),
+                                const SizedBox(
+                                  height: 24,
+                                ),
+                                Text(
+                                    widget.day == "3"
+                                        ? StringUtils.trainingDaysSuccessTitle
+                                        : StringUtils.trainingSuccessTitle,
+                                    style: const TextStyle(
+                                        fontSize: 17, fontWeight: FontWeight.w500)),
+                                const SizedBox(
+                                  height: 20,
+                                ),
+                                widget.day == "1"
                                     ? const Text(
-                                        StringUtils.trainingDay2CompleteSuccess,
+                                        StringUtils.trainingDay1CompleteSuccess,
                                         style: TextStyle(
                                             fontSize: 14,
                                             fontWeight: FontWeight.w300))
-                                    : RichText(
-                                        text: TextSpan(children: <TextSpan>[
-                                          const TextSpan(
-                                              text: StringUtils
-                                                  .trainingDay3CompleteSuccess1,
-                                              style: TextStyle(
-                                                  fontSize: 14,
-                                                  fontWeight: FontWeight.w300)),
-                                          TextSpan(
-                                              text: StringUtils
-                                                  .trainingDay3CompleteSuccess2,
-                                              style: TextStyle(
-                                                  fontSize: 14,
-                                                  color: AppTheme.of(context)
-                                                      .primaryColor,
-                                                  fontWeight: FontWeight.w300)),
-                                          const TextSpan(
-                                              text: StringUtils
-                                                  .trainingDay3CompleteSuccess3,
-                                              style: TextStyle(
-                                                  fontSize: 14,
-                                                  fontWeight: FontWeight.w300)),
-                                        ]),
-                                        textAlign: TextAlign.center),
-                          ],
+                                    : widget.day == "2"
+                                        ? const Text(
+                                            StringUtils.trainingDay2CompleteSuccess,
+                                            style: TextStyle(
+                                                fontSize: 14,
+                                                fontWeight: FontWeight.w300))
+                                        : RichText(
+                                            text: TextSpan(children: <TextSpan>[
+                                              const TextSpan(
+                                                  text: StringUtils
+                                                      .trainingDay3CompleteSuccess1,
+                                                  style: TextStyle(
+                                                      fontSize: 14,
+                                                      fontWeight: FontWeight.w300)),
+                                              TextSpan(
+                                                  text: StringUtils
+                                                      .trainingDay3CompleteSuccess2,
+                                                  style: TextStyle(
+                                                      fontSize: 14,
+                                                      color: AppTheme.of(context)
+                                                          .primaryColor,
+                                                      fontWeight: FontWeight.w300)),
+                                              const TextSpan(
+                                                  text: StringUtils
+                                                      .trainingDay3CompleteSuccess3,
+                                                  style: TextStyle(
+                                                      fontSize: 14,
+                                                      fontWeight: FontWeight.w300)),
+                                            ]),
+                                            textAlign: TextAlign.center),
+                              ],
+                            ),
+                          ),
                         ),
-                      ),
-                    )
+                        (widget.day == "3")?
+                        Center(
+                          //padding: const EdgeInsets.all(16),
+                          child: PinkBorderButton(
+                            isEnabled: true,
+                            content: "Start Exam",
+                            onPressed: () {
+                              Navigator.pushReplacement(
+                                context,
+                                MaterialPageRoute(builder: (context) {
+                                  return TrainingExam(
+                                      title: widget.title, pospId: widget.pospId,);
+                                }),
+                              );
+                            },
+                          ),
+                        ):
+                        Container(),
+                      ],
+                    ),
                   ],
                 ),
               ),
